Activityアニメーションの追加
