+++ app/src/main/AndroidManifest.xml
@@ -9,11 +9,10 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/AppTheme"
+        android:theme="@style/Animation"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
         <activity android:name=".RequestActivity" />
         <activity android:name=".EditActivity" />
-        <activity android:name=".SubActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
