+++ app/src/main/java/com/socu/enpit/sssforshop/EditActivity.kt
@@ -1,12 +1,37 @@
 package com.socu.enpit.sssforshop
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuItem
 
 class EditActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_edit)
+    }
+
+    // メニューを表示させる処理
+    // この関数をオーバーライドして「menu.xml」を指定することで表示される
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.menu, menu)
+        return true
+    }
+
+    // メニューを選択したときの動作をここに書く
+    override fun onOptionsItemSelected(item: MenuItem?): Boolean {
+        val itemId = item?.itemId
+        when (itemId) {
+            // メニューの「ログアウト」を押したとき
+            R.id.menu_logout -> {
+                // MainActivity（ログイン画面）に遷移する
+                val intent = Intent(this, MainActivity::class.java)
+                intent.flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
+                startActivity(intent)
+            }
+        }
+        return super.onOptionsItemSelected(item)
     }
 }
