+++ app/src/main/res/anim/close_enter.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate android:fromXDelta="-100%"
+        android:toXDelta="0%"
+        android:duration="300"
+        android:fillAfter="true"
+        android:fillEnabled="true"/>
+</set>
