+++ app/src/main/res/anim/open_exit.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<set xmlns:android="http://schemas.android.com/apk/res/android">
+    <translate android:fromXDelta="0%"
+        android:toXDelta="-100%"
+        android:duration="300"
+        android:fillAfter="true"
+        android:fillEnabled="true"/>
+</set>
